pyproject.toml
[tool.poetry]
name = "plate-scraper"
version = "0.2.0"
description = "scrape sodexomyway for menu data because their site is so slow"
authors = ["ryleu <69326171+ryleu@users.noreply.github.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
packages = [{include = "plate_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
discord-py-interactions = "^5.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
